requirements.txt
# Intuition-2025 Requirements
# Python dependencies for the Change Management Chat Assistant

# Core Framework
streamlit>=1.29.0

# LangChain Components
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0

# LangGraph for workflow orchestration
langgraph>=0.0.20

# Vector Store & Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Web Search
tavily-python>=0.3.0

# Text Processing
beautifulsoup4>=4.12.0
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Processing (for notebooks)
pandas>=2.0.0
numpy>=1.24.0

# ML & NLP (for notebooks)
scikit-learn>=1.3.0
transformers>=4.35.0

# Web Scraping
googlesearch-python>=1.2.3

# Optional: For notebook analysis
kaggle>=1.5.16
jupyter>=1.0.0
ipykernel>=6.25.0
